Rename comunity to community
