+++ client-web/src/functional-e2e/templates/template-types/verify/community-guidelines-template-verify.ts
